I created the initial js file
I was wondering if I should call it index or script. I have no idea how
I'm going to start this, so I might very well write some pseudocode in a
text file. It's late so I'm definitely not going to start right now. I
would like to have an idea of where I will be taking this soon though.
For example, I want to get used to webpack and using node in general,
but it was a hassle last time and I don't really want to go through it
again. But it is a good technology to know, so I know I will give in
eventually and use it. Let's get to it!