Started designing the UI and Colors
I'm making several html files with static info so I can get the design
down for when I render the parts dynamically. When I'm done designing,
Ill convert the html to js DOM rendering by creating elements and adding
the classes. I find it easier to do it with html first. I'm wondering
how to render the input form for making and editing the todos. I want it
to be right there on the project's page although it would be easier to
make it take the user to a new page for that todo. Personally tho, I
would prefer a site that let me just edit my todos right there on that
page. So I'll look into an edit button that actually expands that
todo div and inserts the form and textboxes etc in there. But then I
would need a cancel button that somehow undoes all of that. And a submit
button that I guess reloads the Project page, or just re-renders the
preview version of that todo. This will be a challenge.