Decided on the UI using html and static info
I'm about finished with the look of the app, and it should be time to
start writing the js files for rendering the pages. I've also been
wrestling with webpack a lot. I watched a tutorial and it helped a
little bit. I hope transferring it from html to js isn't too hard. I'm
probably going to transfer it exactly, keeping the static info, and then
properly make it dynamic and finish this project off.