Finished Rendering the home page
The home page is now dynamic. It took a while but it wasn't as stressful
as I was expecting. The render function for it takes the AppArray (the
array with the projects) because it displays the projects. Currently it
allows you to create a new project with a name and a description.

One thing that really surprises me is that I can import the renderHome
function and when it calls other functions that I haven't imported, it
still works. I never expected that. For example, I simplified the
renderHome function by putting parts of its render process in separate
designated functions, and they still work when I import renderHome into
index.js and call it. I expected it not to work because I didn't import
those other functions that the one I import depends on.

Anyway now it's on to the project page, the page that comes up when you
click on a project from the home page. I'm guessing I will need to pass
the project object to that render function. Also I'll need to add an
option to edit the project from that page. As in editing the project
title and description. I'll probably have to add buttons to the todos as
well. A button to edit, to delete, and to complete, or toggle
completion.