Switched rendering style and got addTodo working
For the renderProject function, I decided to do the rendering
differently. Instead of creating and appending each element, I return a
template string with the html in it and assign it to the innerHTML
property of the container. It's a lot more convenient. I also started
coding the buttons, the first one being the button that adds a todo to
your project. I'm really happy it's working so far.