Can now create a todo without a project
I programmed the create todo button, and had it send the user to the
default project's page with the addTodo form already open. I added an
optional parameter to the renderProject function that if true shows the
addTodo form on launch. I felt good when I did it. I set it to 0 and had
the create todo button eventListener set it to 1. I could have done it
with bools instead of 0 and 1, but I've been practising C recently and
this was muscle memory. I don't think I'll change it. I think this is
the first time I've used an optional parameter by myself and I feel
really resourceful :).