Showed all info in todo div
Before, only the todo title and desc was shown in its div on the Project
page. I added a span that shows the priority and due date as well. I use
the mdn docs to see how I could capitalize the priority and it worked,
using the array slice function. That made me feel really resourceful
too. I think it was in the last commit but I have the complete button
change to incomplete depending on if the todo is actually complete or
not. That was really satisfying. So now I just need to program that
button with an event listener and all three buttons on the todo divs
will be functional. I'm nearing the end of this project. It's been cool
but I really want to move on to react and backend soon. However I don't
want to skip anything important or miss fundamentals either, so I'm just
going along. I wonder if I'll ever add a UI to that tic tac toe site.
Maybe one day. Anyway, the next move is to make the complete button work.