Date problems fixed + Deletion confirmation
I had this problem where the date the user inputted in the date picker
would show up on the site as one day behind. It was caused by this weird
default behaviour of the input[type = "date"] element in html. I'll
actually explain the whole thing here. I was using the .valueAsDate
property of the date picker to both retrieve its info and set its info
(when populating the edit form). What that does is give and take Date
object values, as opposed to just the .value property which gives and
takes date STRINGS ("2003-06-18"). What I found out though, was that
when you select a date, the .valueAsDate Date object that is returned is
set to UTC midnight. That is, prime meridian midnight. I am at UTC -5,
so that means I'm on the previous day at 19:00. That's why when the user
selected a date, it showed up as the day before. The work around, which
I found and find really random, is to use a date string and attach a
time value to it. Instead of .valueAsDate, use .value and retrieve date
strings instead of real date objects. When you retrieve the string add
"T23:59:00" or "T00:00:00" etc, and that will prevent html from assuming UTC midnight. Then make a date object off of the date string using the constructor and store that in the todo.
Aside from that, I searched up how to make an are you sure pop up and found the confirm() function. So I'm using it for deletion. Cool.
What's next is a delete button for projects, ranking todos by priority and local storage.