Renamed dist folder to docs for github pages
Github apparently needs an index.html to be in the root for pages to
work, so I had to find a work around. Especially since the dist folder
was in gitignore. Hopefully this works. I'm publishing it on pages
before I do the local storage but it will be done, trust.