Fixed wrapping problem.
I found a great css value called 'overflow-wrap: break-word;'. It
basically breaks a word if that's the only way to wrap overflowing text.
It worked on the project container on the home page, so I applied it to the
whole document. I found it from an article on tutorialpedia.org called
'CSS Wrap Text: A Comprehensive Guide from MDN'.