Starting to add localStorage
I've started the process of adding local storage functionality to this
todo list as its final step. I'm realising it would be a lot easier to
do the json conversion if I remove the private variables in the todo
class. I was hoping I wouldn't have to do this, but I will. It seems I
can just run the stringify function on the whole app array of projects,
and it will stringify each member of the array and each todo in each
array. What a convenience. I'm going to make the private properties
public for todos but leave the info object there so I won't have to
change anything on the pages hopefully. Let's see. If this works out, I
should be done with this soon!