Re add methods after retrieval from local storage
I figured I would do this by writing static functions in the todo and
project classes that take in a plain version and return a proper object
with methods. Not too hard, but we'll see where I end up putting that
retrieval function. I don't want to have many more modules for this, as
it isn't worth it in my opinion.